scripts/requirements.txt
beautifulsoup4==4.12.2
geopy==2.3.0
hugchat==0.3.7
mock==5.1.0
python-decouple==3.8
requests==2.31.0
selenium==4.12.0
llama_cpp_python==0.2.11
